requirements.txt
# Usage: pip install -r requirements.txt

numpy==1.24.4
pandas==2.0.3
plotly==5.15.0
scikit-learn==1.3.0
scikit-surprise==1.1.3
